+++ pipelines/iot_analytics/create_publish_and_model_files/pubsub_sample.jsonl
